fix : UI 페이지단위 분리 완료
